Auto stash before merge of "master" and "origin/master"

+++ DV1572/DV1572/ourEngine/interface/Window.cpp
@@ -85,7 +85,7 @@ void DX::submitToInstance(Shape* shape, std::vector<DX::INSTANCE_GROUP>& queue)
 	attribDesc.w4 = rows[3];
 
 	attribDesc.highLightColor = shape->getColor(); //This allowes us to use a "click highlight"
-	attribDesc.lightIndex = shape->getLightIndex();
+	attribDesc.lightIndex = static_cast<float>(shape->getLightIndex());
 	
 	// Unique Mesh
 	if (existingId == -1)
@@ -931,7 +931,7 @@ void Window::_geometryPass(const Camera &cam)
 
 		DirectX::XMMATRIX vp = DirectX::XMMatrixTranspose(viewProj);
 		DirectX::XMStoreFloat4x4A(&meshBuffer.VP, vp);
-		meshBuffer.gridscale = instance.shape->getGridScale();
+		meshBuffer.gridscale = static_cast<float>(instance.shape->getGridScale());
 
 		D3D11_MAPPED_SUBRESOURCE dataPtr;
 		DX::g_deviceContext->Map(m_meshConstantBuffer, 0, D3D11_MAP_WRITE_DISCARD, 0, &dataPtr);
@@ -1091,17 +1091,17 @@ void Window::_lightPass(Camera& cam /*std::vector<Light*> lightQueue*/)
 	
 	//Send lights to GPU
 	POINT_LIGHT_COLLECTION pointLightCollectionBuffer; 
-	int nrOfLights = DX::g_lightQueue.size();
+	int nrOfLights = static_cast<int>(DX::g_lightQueue.size());
 
 	for (int i = 0; i < nrOfLights; i++)
 	{
 		pointLightCollectionBuffer.positionArray[i] = DX::g_lightQueue[i]->getPosition();
 		pointLightCollectionBuffer.colorArray[i] =	DX::g_lightQueue[i]->getColor(); 
-		pointLightCollectionBuffer.colorArray[i].w = DX::g_lightQueue[i]->getIndex();
+		pointLightCollectionBuffer.colorArray[i].w = static_cast<float>(DX::g_lightQueue[i]->getIndex());
 		pointLightCollectionBuffer.lightSetup[i] = DX::g_lightQueue[i]->getLightSetup();
 
 	}
-	pointLightCollectionBuffer.nrOfLights = XMFLOAT4A(nrOfLights, nrOfLights, nrOfLights, nrOfLights); 
+	pointLightCollectionBuffer.nrOfLights = XMFLOAT4A(static_cast<float>(nrOfLights), static_cast<float>(nrOfLights), static_cast<float>(nrOfLights), static_cast<float>(nrOfLights));
 
 	D3D11_MAPPED_SUBRESOURCE lightData;
 	DX::g_deviceContext->Map(m_pPointLightBuffer, 0, D3D11_MAP_WRITE_DISCARD, 0, &lightData);
