+++ DV1572/DV1572/ourEngine/interface/shape/Mesh.cpp
@@ -366,7 +366,7 @@ int Mesh::getNrOfIndices(int i) const
 
 int Mesh::getNumberOfParts() const
 {
-	return m_materials.size();
+	return static_cast<int>(m_materials.size());
 }
 
 bool Mesh::CheckID(const Mesh& other) const
